+++ software_esp32/include/compile_time.h
@@ -78,5 +78,3 @@
 
 #endif /* COMPILE_TIME_H_ */
 
-
-