+++ software_esp32/include/helper.h
@@ -51,3 +51,6 @@ bool isNumber(const String& str);
 bool isFloat(const std::string& str);
 void replace (char* buffer,uint16_t size,char find, char with);
 uint32_t versionExplode (String sv);
+char* findCharInString (char c,char* pString,size_t size);
+char* copyStringUntilChar (char* pString,char* buffer,char c,size_t pSize,size_t bSize);
+size_t strlcat(char *dst, const char *src, size_t dsize);