+++ software_esp32/src/helper.cpp
@@ -112,3 +112,55 @@ uint32_t versionExplode (String sv)
   // UART_DBG.println("versionExplode "+sv+": result="+String(result,16));
   return result;
 }
+
+ char* findCharInString (char c,char* pString,size_t size)
+ {
+  char* p = pString;
+  size_t len = 0;
+  while ((*p!=c) && (*p!=0) && (len<size)) {
+    p++;
+    len++;
+  }
+  if ((*p==0) || (len==size)) p=NULL;
+  return (p);
+ }
+
+ char* copyStringUntilChar (char* pString,char* buffer,char c,size_t pSize,size_t bSize)
+ {
+  char* p = pString;
+  char* b = buffer;
+  size_t len = 0;
+  while ((*p!=c) && (*p!=0) && (len<pSize) && (len<bSize)) {
+    *b++=*p++;
+    len++;
+  }
+  if (len>=bSize) return NULL;
+  return p;
+ }
+
+size_t strlcat(char *dst, const char *src, size_t dsize)
+{
+	const char *odst = dst;
+	const char *osrc = src;
+	size_t n = dsize;
+	size_t dlen;
+
+	/* Find the end of dst and adjust bytes left but don't go past end. */
+	while (n-- != 0 && *dst != '\0')
+		dst++;
+	dlen = dst - odst;
+	n = dsize - dlen;
+
+	if (n-- == 0)
+		return(dlen + strlen(src));
+	while (*src != '\0') {
+		if (n != 0) {
+			*dst++ = *src;
+			n--;
+		}
+		src++;
+	}
+	*dst = '\0';
+
+	return(dlen + (src - osrc));	/* count does not include NUL */
+}